requirements.txt
gym == 0.23.1
numpy < 2
jax[cuda12]==0.4.35
distrax==0.1.5
flax==0.10.2
ml_collections >= 0.1.0
tqdm >= 4.60.0
chex==0.1.87
optax==0.2.4
orbax-checkpoint==0.10.1
absl-py >= 0.12.0
scipy >= 1.6.0
wandb >= 0.12.14
plotly
tensorflow == 2.18
tensorflow-probability[tf] == 0.25.0
einops >= 0.6.1
imageio >= 2.31.1
moviepy >= 1.0.3
pre-commit == 3.3.3
opencv-python
seaborn
omegaconf
overrides
transformers
d4rl @ git+https://github.com/maxsobolmark/D4RL
cython < 3
patchelf
dm_control <= 1.0.20
mujoco <= 3.1.6
hydra-core
